requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymysql==1.1.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
